+++ Pulsar4X/Pulsar4X.ECSLib/ComponentFeatureSets/MineResources/MineingDB.cs
@@ -5,13 +5,13 @@ namespace Pulsar4X.ECSLib
 {
     public class MiningDB : BaseDataBlob
     {
-        public Dictionary<Guid, int> MineingRate { get; set; }
+        public Dictionary<Guid, long> MineingRate { get; set; }
 
         public Dictionary<Guid, MineralDepositInfo> MineralDeposit => OwningEntity.GetDataBlob<ColonyInfoDB>().PlanetEntity.GetDataBlob<SystemBodyInfoDB>().Minerals;
 
         public MiningDB()
         {
-            MineingRate = new Dictionary<Guid, int>();
+            MineingRate = new Dictionary<Guid, long>();
         }
 
         public MiningDB(MiningDB db)
